requirements.txt
django==2.2.7
djangorestframework==3.10.3
drf-yasg==1.17.0
django-admin==2.0.0
python-dotenv==0.10.3
boto3==1.10.26
django-storages==1.8
gunicorn==20.0.2
django-rest-auth==0.9.5
django-allauth==0.40.0
flake8